+++ apps/playground/rasengan-io-demo/src/main.tsx
@@ -3,7 +3,7 @@ import { RasenganIOProvider } from '@rasenganjs/io';
 
 export default function App({ Component, children }: AppProps) {
   return (
-    <RasenganIOProvider url="http://localhost:4000" autoConnect={true}>
+    <RasenganIOProvider url="ws://localhost:4000/lobby" autoConnect={true}>
       <Component>{children}</Component>
     </RasenganIOProvider>
   );